requirements.txt
asgiref==3.7.2
distlib==0.3.6
Django==4.2.2
django-ckeditor==6.6.1
django-environ==0.10.0
django-js-asset==2.1.0
filelock==3.12.2
gunicorn==20.1.0
Pillow==9.5.0
platformdirs==3.8.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
sqlparse==0.4.4
typing_extensions==4.6.3
virtualenv==20.23.1
